Make updater less complex
