+++ src/main/java/net/skinsrestorer/bukkit/utils/UpdateDownloaderGithub.java
@@ -22,12 +22,18 @@
 package net.skinsrestorer.bukkit.utils;
 
 import net.skinsrestorer.bukkit.SkinsRestorer;
+import net.skinsrestorer.shared.exception.UpdateException;
 import net.skinsrestorer.shared.update.DownloadCallback;
 import net.skinsrestorer.shared.update.GitHubReleaseInfo;
-import net.skinsrestorer.shared.update.GitHubUpdateDownloader;
 import org.bukkit.Bukkit;
 
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.nio.channels.Channels;
+import java.nio.channels.ReadableByteChannel;
 import java.util.logging.Level;
 
 public class UpdateDownloaderGithub extends UpdateDownloader {
@@ -64,7 +70,7 @@ public class UpdateDownloaderGithub extends UpdateDownloader {
         final File updateFile = new File(updateFolder, pluginFile.getName());
 
         plugin.getLogger().info("[GitHubUpdate] Downloading update...");
-        Bukkit.getScheduler().runTaskAsynchronously(plugin, GitHubUpdateDownloader.downloadAsync(releaseInfo, updateFile, plugin.getUpdateChecker().getUserAgent(), new DownloadCallback() {
+        Bukkit.getScheduler().runTaskAsynchronously(plugin, downloadAsync(releaseInfo, updateFile, plugin.getUpdateChecker().getUserAgent(), new DownloadCallback() {
             @Override
             public void finished() {
                 plugin.getLogger().info("[GitHubUpdate] Update saved as " + updateFile.getPath());
@@ -77,5 +83,41 @@ public class UpdateDownloaderGithub extends UpdateDownloader {
         }));
 
         return true;
+    }
+
+    private static Runnable downloadAsync(final GitHubReleaseInfo releaseInfo, final File file, final String userAgent, final DownloadCallback callback) {
+        return () -> {
+            try {
+                download(releaseInfo, file, userAgent);
+                callback.finished();
+            } catch (Exception e) {
+                callback.error(e);
+            }
+        };
+    }
+
+    private static void download(GitHubReleaseInfo releaseInfo, File file, String userAgent) {
+        ReadableByteChannel channel;
+
+        try {
+            HttpURLConnection connection = (HttpURLConnection) new URL(releaseInfo.latestDownloadURL).openConnection();
+            connection.setRequestProperty("User-Agent", userAgent);
+            if (connection.getResponseCode() != 200) {
+                throw new UpdateException("Download returned status #" + connection.getResponseCode());
+            }
+
+            channel = Channels.newChannel(connection.getInputStream());
+        } catch (IOException e) {
+            throw new UpdateException("Download failed", e);
+        }
+
+        try {
+            FileOutputStream output = new FileOutputStream(file);
+            output.getChannel().transferFrom(channel, 0L, 9223372036854775807L);
+            output.flush();
+            output.close();
+        } catch (IOException e) {
+            throw new UpdateException("Could not save file", e);
+        }
     }
 }
